requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.4.0
google-generativeai>=0.6.0
pdfplumber>=0.8.0
python-docx>=0.8.11
requests>=2.28.0
pydantic>=1.10.0
fastapi
uvicorn
python-dotenv
requests
sqlalchemy
psycopg2-binary
pgvector
